TIP-1223: change the naming of indexable product
