+++ src/Akeneo/Pim/Enrichment/Bundle/Elasticsearch/GetElasticsearchProductProjectionInterface.php
@@ -5,15 +5,22 @@ declare(strict_types=1);
 namespace Akeneo\Pim\Enrichment\Bundle\Elasticsearch;
 
 use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Model\ElasticsearchProductProjection;
+use Akeneo\Pim\Enrichment\Component\Product\Exception\ObjectNotFoundException;
 
 /**
  * @author    Nicolas Marniesse <nicolas.marniesse@akeneo.com>
  * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface GetIndexableProductInterface
+interface GetElasticsearchProductProjectionInterface
 {
+    /**
+     * @throws ObjectNotFoundException when the product is not found
+     */
     public function fromProductIdentifier(string $productIdentifier): ElasticsearchProductProjection;
 
+    /**
+     * @throws ObjectNotFoundException when one or several of the product are not found
+     */
     public function fromProductIdentifiers(array $productIdentifiers): array;
 }
