+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/queries.yml
@@ -245,8 +245,8 @@ services:
         arguments:
             - '@database_connection'
 
-    akeneo.pim.enrichment.product_model.query.get_indexable_product:
-        class: Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Query\GetIndexableProduct
+    akeneo.pim.enrichment.product.query.get_elasticsearch_product_projection:
+        class: Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Query\GetElasticsearchProductProjection
         arguments:
             - '@pim_catalog.repository.product'
             - '@pim_catalog.repository.locale'
