+++ tests/back/Pim/Enrichment/Specification/Bundle/Elasticsearch/Indexer/ProductIndexerSpec.php
@@ -3,7 +3,7 @@
 namespace Specification\Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Indexer;
 
 use Akeneo\Pim\Enrichment\Component\Product\Connector\ReadModel\IndexableProduct;
-use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\GetIndexableProductInterface;
+use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\GetElasticsearchProductProjectionInterface;
 use Akeneo\Tool\Bundle\ElasticsearchBundle\Client;
 use Akeneo\Tool\Bundle\ElasticsearchBundle\Refresh;
 use Akeneo\Tool\Component\StorageUtils\Indexer\ProductIndexerInterface;
@@ -15,7 +15,7 @@ class ProductIndexerSpec extends ObjectBehavior
 {
     function let(
         Client $productAndProductModelIndexClient,
-        GetIndexableProductInterface $getIndexableProduct
+        GetElasticsearchProductProjectionInterface $getIndexableProduct
     ) {
         $this->beConstructedWith($productAndProductModelIndexClient, $getIndexableProduct);
     }
